Expose pino.pretty()
